+++ addons/mail/static/src/models/attachment.js
@@ -77,18 +77,23 @@ registerModel({
             if (!this.isUploading) {
                 this.update({ isUnlinkPending: true });
                 try {
-                    await this.async(() => this.messaging.rpc({
+                    await this.messaging.rpc({
                         route: `/mail/attachment/delete`,
                         params: {
                             access_token: this.accessToken,
                             attachment_id: this.id,
                         },
-                    }, { shadow: true }));
+                    }, { shadow: true });
                 } finally {
-                    this.update({ isUnlinkPending: false });
+                    if (this.exists()) {
+                        this.update({ isUnlinkPending: false });
+                    }
                 }
             } else if (this.uploadingAbortController) {
                 this.uploadingAbortController.abort();
+            }
+            if (!this.exists()) {
+                return;
             }
             this.delete();
         },
